117. Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -27,9 +27,11 @@ const storage = multer.diskStorage({
 });
 
 router.post(
-    "", checkAuth, multer({storage: storage}).single("image"),
+    "", 
+    checkAuth,
+    multer({storage: storage}).single("image"),
  (req, res, next) => {
-    const url = req.protocol + '://' + req.get("host");
+    const url = req.protocol + "://"+ req.get("host");
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
@@ -48,12 +50,15 @@ router.post(
     });
 });
 
-router.put("/:id", checkAuth, multer({ storage: storage }).single("image"),
-(req, res, next) => {
-    let imagePath = req.body.imagePath;
+router.put(
+ "/:id",
+  checkAuth,
+  multer({ storage: storage }).single("image"),
+  (req, res, next) => {
+   let imagePath = req.body.imagePath;
    if (req.file) {
        const url = req.protocol + "://" + req.get("host");
-       imagePath = url + "/images/" + req.file.filename
+       imagePath = url + "/images/" + req.file.filename;
    }
     const post = new Post({
       _id: req.body.id,
@@ -61,9 +66,13 @@ router.put("/:id", checkAuth, multer({ storage: storage }).single("image"),
       content: req.body.content,
       imagePath: imagePath
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if (result.n > 0) {
+            res.status(200).json({ message: "Update successful!" });
+        } else {
+            res.status(401).json({ message: "Not authorized"});
+        }
+     
     });
   });
 
@@ -101,9 +110,13 @@ router.get("/:id", (req, res, next) => {
 
 
     router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then(result => {
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
         console.log(result);
-        res.status(200).json({ message: "Post deleted!" });
+        if (result.n > 0) {
+            res.status(200).json({ message: "Deletion successful!" });
+        } else {
+            res.status(401).json({ message: "Not authorized"});
+        }
     });
 
 });
